avdevice/dshow: Add support for capturing H264 (FORMAT_UVCH264Video) from webcams
The Logitech C930e webcam (and possibly others) present their H264 capabilities through a KS_H264VIDEOINFO structure (format type FORMAT_UVCH264Video) rather than through the VIDEOINFOHEADER or VIDEOINFOHEADER2 structures used for other formats.
This commit adds support to dshow (dshow.c) to allow these formats to be listed and captured.
The GUIDs required for this along with the KS_H264VIDEOINFO structure are also included (dshow_capture.h) as they do not appear to be in the MingGW headers.

TODO: Add support to choose the H264 profile used. The C930e seems to present both a Constrained Baseline and a High Profile + 0x0C. Currently this is ignored and the first one matching the requested resolution is used.