avdevice/dshow: Add support for specifying the H264 profile to capture
Previous commit did not allow the user to specify which H264 profile to capture (used first profile with correct resolution and framerate).
This commit allows the user to specify which H264 profile to try to capture using the -h264_profile argument.
Possible parameters for the -h264_profile argument are baseline, main and high.